frontend/requirements.txt
Flask==3.1.0
authlib==1.4.0
google-cloud-compute==1.22.0
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.2
